Fix Independent Enrollment (#268)
# Fix Independent Enrollment

## :recycle: Current situation & Problem
We were notified about the following issues:
- When the patient does the registration survey (and updating data
survey), the survey only asks if the patient is taking a class of
medication (ex: beta blocker, MRA) but does not ask any specific
questions as to which medication or dosage they are on. And after
completing either surveys, there are still no medications uploaded to
the patients account so none appear on the app or on the health summary.
_--> see: https://github.com/StanfordBDHG/ResearchKitOnFHIR/pull/96_

- Similarly for the upcoming appointments on the registration survey and
updating data survey, the dates that are inputted for the upcoming
appointments are not saved to the patient account.
_--> this is actually a result of the errors from the response services
failing while evaluating the questionnaire responses_

- After completing the registration survey, the notification to complete
the registration survey does not disappear and allows the patient to
submit it again.
_--> this is actually a result of the errors from the response services
failing while evaluating the questionnaire responses_

- When adding lab values on the registration survey and update data
survey, the date that is saved on the survey for when the patient had
the measurement taken is different from what is saved on the dashboard.
It seems to be set as whatever date the patient inputs, saves as the day
before instead. For ex: If I inputted that I took a creatine measurement
of 1.2 today 7/8/2025, on the dashboard it would say that i have a
creatine measurement of 1.2 on 7/7/2025.
_--> this is actually an issue where we are getting a date without time
information and then we store it with time information, which doesn't
really make sense._

## :gear: Release Notes
*Add a bullet point list summary of the feature and possible migration
guides if this is a breaking change so this section can be added to the
release notes.*
*Include code snippets that provide examples of the feature implemented
or links to the documentation if it appends or changes the public
interface.*


## :books: Documentation
*Please ensure that you properly document any additions in conformance
to [Spezi Documentation
Guide](https://github.com/StanfordSpezi/.github/blob/main/DOCUMENTATIONGUIDE.md).*
*You can use this section to describe your solution, but we encourage
contributors to document your reasoning and changes using in-line
documentation.*


## :white_check_mark: Testing
*Please ensure that the PR meets the testing requirements set by CodeCov
and that new functionality is appropriately tested.*
*This section describes important information about the tests and why
some elements might not be testable.*


### Code of Conduct & Contributing Guidelines
By creating and submitting this pull request, you agree to follow our
[Code of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md).